Rename `type_def` to `typedef` for consistency with `opdef`.
