+++ src/resource/typedef.rs
